Adding special votes (#198)
Creates a window to choose the type of special vote
- Adding the argument "playerForSpecialVote"
- Updating the history to take account of the special votes
- Allowing all players to vote during special vote
- Adding an option "Special vote" in players' menu
NB: This skull will not be visible while the vote is still open. You need to close the vote first.
If the vote is detected as special, the majority isn't printed

This change affects the Story Teller's mark (for the case of an Atheist-script, where the Story Teller is marked as being on the block).

This mark is now hidden for players if the option "Organ Grinder Vote" is turned on.