Updating the Organ Grinder Vote Mode (#205)
* Update CHANGELOG.md

* Hidding the skulls for player, if the option "Organ Grinder Vote" is ON.

* Adding a message for the ST, if the option "Organ Grinder Vote" is ON.

* Correcting spaces typos

* Adding the french version of the message for ST

* Adding the english version of the message for the ST

* Deleting an OG reminder (EN)

This reminder was previously used to ignore the issue of the skulls visible by players. Now that this issue is corrected, this token is useless.

* Deleting an OG reminder (FR)

This reminder was previously used to ignore the issue of the skulls visible by players. Now that this issue is corrected, this token is useless.

* Some corrections for the linter

* Some corrections for the linter

* Corrections for the linter

---------

Co-authored-by: Pingumask <68610022+Pingumask@users.noreply.github.com>